+++ Sources/MapboxMaps/Foundation/MapViewDependencyProvider.swift
@@ -1,4 +1,3 @@
-import UIKit
 import MetalKit
 
 internal protocol MapViewDependencyProviderProtocol: AnyObject {