+++ Tests/MapboxMapsTests/Foundation/Mocks/MockDisplayLink.swift
@@ -1,6 +1,5 @@
-import Foundation
+import UIKit
 @testable import MapboxMaps
-import QuartzCore
 
 final class MockDisplayLink: CADisplayLink {